+++ src/Symfony/Component/Intl/CHANGELOG.md
@@ -14,3 +14,5 @@ CHANGELOG
    names when fallback (e.g. "en_GB"->"en") is possible anyway. This reduced
    the Resources/ directory file size of the Icu 1.2.x branch from 14M to 12M at
    the time of this writing
+ * `StructuredBundleReader` now follows aliases when looking for fallback locales
+ * [BC BREAK] a new method `getLocaleAliases()` was added to `LocaleBundleInterface`